Release RT fixes using EUMM
Make final modifications to the build files so that we can release
a version containing just the RT fixes
(and as few other changes as possible
(in the hopes of fixing rather than breaking)).

Hopefully future releases will be done with Dist::Zilla
(and all these files will be generated automatically).