Remove unnecessary :s/// command on source file
There's no reason to modify the input file
(and in this case the effect was completely unused)..

This substitution command appears to have been cargo-culted
(incorrectly) from the original 2html.vim script.

If the file in fact had any end-of-line whitespace
vim would throw an error
(since Text::VimColor uses -R to set 'readonly')
which would cause a 1 second delay.

Closes gh-2.